✅ ::  divide Activity and Screen
2023-01-24
