+++ app/src/main/java/kr/hs/dgsw/smartschool/dui/root/MainActivity.kt
@@ -125,114 +125,6 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Preview(showBackground = true)
-@Composable
-fun TestPreview(){
-    Surface(
-        modifier = Modifier.fillMaxSize()
-    ){
-        MainScreen(viewModel = null)  
-    }
-}
-
-data class Item(
-    val icon : Int,
-    val title : String,
-    val content : String
-    )
-
-
-@Composable
-fun MainScreen(viewModel : ViewModel?){
-    val list : List<Item> = listOf(
-        Item(kr.hs.smartschool.components.R.drawable.ic_breakfast_3d,"제목","내용"),
-        Item(kr.hs.smartschool.components.R.drawable.ic_breakfast_3d,"제목","내용"),
-        Item(kr.hs.smartschool.components.R.drawable.ic_breakfast_3d,"제목","내용")
-    )
-    Surface(
-        modifier = Modifier
-            .fillMaxSize()
-    ) {
-        ListScreen(list = list)
-    }
-}
-@Composable
-fun ListScreen(
-    list : List<Item>
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(start = 10.dp, 10.dp, 10.dp, 10.dp)
-    ) {
-        Title1(
-            text = "DUI",
-            modifier = Modifier
-                .padding(start = 30.dp, bottom = 50.dp)
-        )
-        ColumnList(list = list)
-    }
-}
-
-@Composable
-fun ColumnList(list : List<Item>){
-    LazyColumn(
-        modifier = Modifier
-            .fillMaxSize()
-            .horizontalScroll(rememberScrollState()),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ){
-
-        list.let {
-            items(it) { item ->
-                ColumnItemCard(
-                    item = item
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun ColumnItemCard(item: Item) {
-    Card(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(vertical = 10.dp),
-        shape = MaterialTheme.shapes.small.copy(CornerSize(20.dp)),
-        elevation = 3.dp
-    ) {
-        Column(modifier = Modifier
-            .height(100.dp)
-            .width(300.dp)
-            .padding(all = 10.dp)
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxSize(),
-                verticalAlignment = Alignment.Top
-            ) {
-                ItemImage(
-                    item.icon
-                )
-                Title3(
-                    text = item.title
-                )
-            }
-            Body3(text = "내용")
-        }
-    }
-}
-
-@Composable
-fun ItemImage(icon : Int){
-    Surface(shape = MaterialTheme.shapes.small.copy(CornerSize(20.dp))) {
-        Image(
-            painter = painterResource(id = icon),
-            contentDescription = null
-        )
-    }
-}
 
 private fun sampleOnClick(context: Context) {
     Toast.makeText(context, "Hello", Toast.LENGTH_SHORT).show()
